Allowing to return or skip invalid dates
On monthly repetitions, invalid dates like 2020-11-31 can be generated.

The way the lib was working until now was that it would fallback to the
last day of the month in that case, but as the RFC for RRULE has a
different take on that, I think it is a nice to have, at least as an
option.

So I added the `:invalid` option to the rules, allowing the devs either
get the invalid dates by setting it to `:return` or skipping them by
setting to `:skip`. Defaults to `:fallback`